requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# LLM and AI
google-generativeai==0.3.1

# Vector database and embeddings
chromadb==0.4.18
sentence-transformers==2.2.2

# Data processing
pandas==2.1.3
openpyxl==3.1.2
numpy==1.26.2

# Web scraping (for data collection - optional)
beautifulsoup4==4.12.2
requests==2.31.0

# MCP Server (optional)
fastmcp==0.1.0
mcp==0.1.0

# Utilities
python-multipart==0.0.6
